pyproject.toml
[project]
name = "langchn"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "langchain>=1.0.7",
    "langchain-google-genai>=3.0.3",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
    "rav>=0.1.1",
]
